readd crowdin script
