+++ builder/src/constants.ts
@@ -1,3 +1,5 @@
 export const sortJsonOptions = {
     indentSize: 4,
 };
+
+export const crowdinProjectId = 528454;
