+++ builder/src/utils/wait.ts
@@ -1,0 +1,1 @@
+export const wait = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
